auth done

+++ next/context/user-context.tsx
@@ -7,7 +7,15 @@ import React, {
   ReactNode,
 } from "react";
 import { User } from "@/types/types";
-import { AuthResult, createAccount, signIn } from "@/lib/strapi/auth";
+import {
+  AuthResult,
+  createAccount,
+  ForgotPasswordResult,
+  makeForgotPasswordRequest,
+  makeResetPasswordRequest,
+  ResetPasswordResult,
+  signIn,
+} from "@/lib/strapi/auth";
 
 const UserContext = createContext<UserContextType | null>(null);
 
@@ -16,6 +24,12 @@ type UserContextType = {
   token: string | null;
   login: (email: string, password: string) => Promise<Partial<AuthResult>>;
   register: (email: string, password: string) => Promise<Partial<AuthResult>>;
+  forgotPassword: (email: string) => Promise<ForgotPasswordResult>;
+  resetPassword: (props: {
+    code: string;
+    password: string;
+    confirmPassword: string;
+  }) => Promise<ResetPasswordResult>;
   logout: () => void;
 };
 
@@ -35,10 +49,23 @@ export const UserProvider = ({ children }: { children: ReactNode }) => {
     }
   }, []);
 
+  const register = async (email: string, password: string) => {
+    if (typeof window !== "undefined") {
+      const { ok, user, token, message } = await createAccount(email, password);
+      if (ok && user && user.confirmed && !user.blocked && token) {
+        localStorage.setItem("token", token);
+        localStorage.setItem("user", JSON.stringify(user));
+        setToken(token);
+        setUser(user);
+      }
+      return { message, ok };
+    }
+  };
+
   const login = async (email: string, password: string) => {
     if (typeof window !== "undefined") {
       const { ok, user, token, message } = await signIn(email, password);
-      if (ok && user && token) {
+      if (ok && user && user.confirmed && !user.blocked && token) {
         localStorage.setItem("token", token);
         localStorage.setItem("user", JSON.stringify(user));
         setToken(token);
@@ -57,21 +84,34 @@ export const UserProvider = ({ children }: { children: ReactNode }) => {
     }
   };
 
-  const register = async (email: string, password: string) => {
-    if (typeof window !== "undefined") {
-      const { ok, user, token, message } = await createAccount(email, password);
-      if (ok && user && token) {
-        localStorage.setItem("token", token);
-        localStorage.setItem("user", JSON.stringify(user));
-        setToken(token);
-        setUser(user);
-      }
-      return { message, ok };
-    }
+  const forgotPassword = async (email: string) => {
+    return await makeForgotPasswordRequest(email);
+  };
+
+  const resetPassword = async ({
+    code,
+    password,
+    confirmPassword,
+  }: {
+    code: string;
+    password: string;
+    confirmPassword: string;
+  }) => {
+    return await makeResetPasswordRequest(code, password, confirmPassword);
   };
 
   return (
-    <UserContext.Provider value={{ user, token, login, register, logout }}>
+    <UserContext.Provider
+      value={{
+        user,
+        token,
+        login,
+        register,
+        logout,
+        forgotPassword,
+        resetPassword,
+      }}
+    >
       {children}
     </UserContext.Provider>
   );
